+++ apps/er-oscoap/opt-cbor.h
@@ -3,7 +3,7 @@
 #include <stddef.h>
 #include <inttypes.h>
 
-uint8_t OPT_CBOR_put_text(uint8_t **buffer, char *text, size_t text_len);
+uint8_t OPT_CBOR_put_text(uint8_t **buffer, char *text, uint8_t text_len);
 
 uint8_t OPT_CBOR_put_array(uint8_t **buffer,uint8_t elements);
